cast: do not downgrade to float when double should be used
Regression from commit 78f019aa9d49be5ada89296a462939738c20c181

Requested-by: rewolff <R.E.Wolff@BitWizard.nl>
Reference: https://github.com/traviscross/mtr/pull/135/files#r74889063

+++ report.c
@@ -472,7 +472,7 @@ void csv_close(time_t now)
 
       /* 1000.0 is a temporay hack for stats usec to ms, impacted net_loss. */
       if( strchr( data_fields[j].format, 'f' ) ) {
-	printf( ",%.2f", (float) (data_fields[j].net_xxx(at) / 1000.0));
+	printf( ",%.2f", (double) (data_fields[j].net_xxx(at) / 1000.0));
       } else {
 	printf( ",%d",   data_fields[j].net_xxx(at) );
       }
